+++ src/main/java/com/example/e_commerce/repository/CategoryRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.e_commerce.model.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
     public Optional<Category> findByName(String name);
 
-    public Optional<Category> findById(Long id);
+    public Optional<Category> findById(Integer id);
 
 }
