+++ src/main/java/com/example/e_commerce/repository/VendorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.e_commerce.model.Vendor;
 
-public interface VendorRepository extends JpaRepository<Vendor, Long> {
+public interface VendorRepository extends JpaRepository<Vendor, Integer> {
     Optional<Vendor> findByEmail(String email);
 
 }
